refactor: remove some unused code, move met data read code out of driver into new load module in met_data
