+++ src/monarchs/core/metadata.py
@@ -1,16 +1,18 @@
 """
 Run metadata for MONARCHS output files.
 
-Two jobs:
+This outputs model_setup parameters as JSON, with the intent of allowing
+the user to re-run a model simulation from this metadata in future -
+e.g. for if you change your model setup script after running. A
+future tool will be able to read this metadata and re-run the model with
+the relevant parameters. It also contains details of the exact model
+version used to generate a given output.
+TODO - actually write this tool
 
-- ``global_attrs(model_setup)``: run-provenance attributes (code version,
-  git state, dependency versions, the fully-resolved configuration) written
-  as netCDF *global* attributes on every output and checkpoint file, so any
-  result file records exactly what produced it (``ncdump -h file.nc``).
-
-- ``VARIABLE_METADATA`` / ``apply_variable_metadata``: units and long_name
-  for model grid fields, applied to each netCDF variable written, so output
-  files are self-describing.
+Additionally, we handle variable metadata here for the output netCDFs.
+This is a WIP - not all fields are covered at time of writing, but the idea
+is to aid reproducibility and introduce a kind of convention for MONARCHS
+output data, somewhat inspired by the CF conventions.
 """
 
 import hashlib
@@ -40,9 +42,9 @@ def global_attrs(model_setup=None):
         except PackageNotFoundError:
             pass
 
-    # Exact git state when running from a checkout (editable install);
-    # silently absent for wheel installs, where the version attribute
-    # already pins the release.
+    # get the git description of the model versioning used. if running
+    # from the pypi install, then it will use the exact release version, else
+    # it will be related to the branch/checkout used.
     try:
         attrs["git_describe"] = subprocess.check_output(
             ["git", "describe", "--tags", "--dirty", "--always"],
@@ -60,12 +62,16 @@ def global_attrs(model_setup=None):
 
 def _summarise_config(model_setup):
     """
-    The fully-resolved configuration (post-defaults) as a JSON-safe dict.
-    Arrays are summarised by shape and content hash rather than dumped.
+    Generate MONARCHS model setup config as a JSON.
+
+    Array values are given as strigns here rather than being dumped -
+    the idea is that we *can* reconstruct this data, not that we necessarily
+    need to save/cache it
     """
     config = {}
     for key, value in sorted(vars(model_setup).items()):
         if isinstance(value, np.ndarray):
+            # digest - we only need a short hash here, just to identify the array in a reproducible way
             digest = hashlib.sha256(np.ascontiguousarray(value)).hexdigest()[:12]
             config[key] = f"<array shape={value.shape} sha256:{digest}>"
         else:
@@ -76,25 +82,25 @@ def _summarise_config(model_setup):
 # units / long_name for model grid fields, keyed by field name.
 # Fractions and flags are dimensionless ("1" per CF convention).
 VARIABLE_METADATA = {
-    "firn_depth": ("m", "Firn column depth"),
+    "firn_depth": ("m", "Firn column total depth"),
     "firn_temperature": ("K", "Firn column temperature"),
     "vertical_profile": ("m", "Depth of each firn layer below the surface"),
-    "rho": ("kg m-3", "Firn bulk density"),
-    "rho_lid": ("kg m-3", "Frozen lid bulk density"),
+    "rho": ("kg m-3", "Firn density"),
+    "rho_lid": ("kg m-3", "Frozen lid density"),
     "Sfrac": ("1", "Solid (ice) volume fraction"),
     "Lfrac": ("1", "Liquid (water) volume fraction"),
-    "albedo": ("1", "Surface albedo"),
+    "albedo": ("1", "Surface albedo, determined by current model state"),
     "meltflag": ("1", "Meltwater present at layer (flag)"),
     "saturation": ("1", "Layer saturated (flag)"),
-    "water": ("m", "Liquid water depth per layer (lateral-flow working field)"),
-    "water_level": ("m", "Hydraulic head used for lateral flow"),
-    "lake_depth": ("m", "Surface lake depth"),
+    "water": ("m", "Liquid water depth per layer, derived from LFrac for firn water"),
+    "water_level": ("m", "Water height for lateral flow"),
+    "lake_depth": ("m", "Melt lake depth"),
     "lake_temperature": ("K", "Lake temperature profile"),
     "lake": ("1", "Lake present (flag)"),
     "lid_depth": ("m", "Frozen lid depth"),
     "lid_temperature": ("K", "Frozen lid temperature profile"),
     "lid": ("1", "Frozen lid present (flag)"),
-    "v_lid": ("1", "Virtual (embryonic) lid present (flag)"),
+    "v_lid": ("1", "Virtual lid present (flag)"),
     "v_lid_depth": ("m", "Virtual lid depth"),
     "virtual_lid_temperature": ("K", "Virtual lid temperature"),
     "ice_lens": ("1", "Ice lens present (flag)"),
@@ -102,19 +108,22 @@ VARIABLE_METADATA = {
     "melt": ("1", "Surface melt occurred this step (flag)"),
     "melt_hours": ("h", "Cumulative hours of surface melt"),
     "exposed_water": ("1", "Exposed water at surface (flag)"),
-    "total_melt": ("m", "Cumulative melt (ice thickness equivalent)"),
-    "lid_sfc_melt": ("m", "Tracked (unremoved) lid surface melt"),
-    "lid_snow_depth": ("m", "Snow depth on the frozen lid"),
+    "total_melt": ("m", "Cumulative melt depth"),
+    "lid_sfc_melt": ("m", "Tracked lid surface melt"),
+    "lid_snow_depth": ("m", "Snow depth on the frozen lid, tracked for albedo changes"),
     "snow_added": ("m", "Snow depth added"),
     "firn_boundary_change": ("m", "Firn boundary change this day"),
     "lake_boundary_change": ("m", "Lake boundary change this day"),
     "lid_boundary_change": ("m", "Lid boundary change this day"),
-    "water_direction": ("1", "Lateral outflow direction flags (0=NW..7=W)"),
+    "water_direction": ("1", "Lateral outflow direction (0=NW..7=W)"),
     "lat": ("degrees_north", "Latitude of grid cell"),
     "lon": ("degrees_east", "Longitude of grid cell"),
     "size_dx": ("m", "Grid cell size, east-west"),
     "size_dy": ("m", "Grid cell size, north-south"),
-    "valid_cell": ("1", "Cell participates in the model physics (flag)"),
+    "valid_cell": (
+        "1",
+        "Cell is included in the set of cells running model physics (flag)",
+    ),
 }
